Add git-post-push-compare-url (#1)
This is a script that allows us to print the URL of the Compare View for
a particular push.

Git doesn't have a post-push hook because of weird pedantic hand wavery
reasons, so instead I'm tying this into `gp`, which is my alias for
pushing to the remote. The porcelain version of the command is used and
piped to `git-post-push-compare-url`, which simply checks to see if it's
a github.com URL and then prints it to stdout. From there I can
triple-click the line to quickly copy it and paste it elsewhere if I
want. I didn't want to go so far as to pipe it to `open` or something,
although that could be a possibility, too.